requirements.txt
PyGithub
boto3